requirements.txt
flask==3.0.0
pymongo==4.6.0
openai==1.6.1
anthropic==0.8.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-anthropic==0.0.1
python-dotenv==1.0.0
numpy==1.24.3
python-dateutil==2.8.2
tqdm==4.66.1